Fix query see BT#12735
The problem.
Course A is accessible on URL 1 and URL 2
Session S i created on URL 2 and include Course A
on the page plugin/buycourses/src/configuration.php on URL 1 the course does not appear.

Expected result

Course A should appear on the page plugin/buycourses/src/configuration.php
on URL 1 and should not appear on URL 2 (because it is included in a session of this URL).