Agenda: Repeated "All day" events show on full day span after first occurence #3394
From the second round, allDay changes from true to 1, as compared to the string true, it will not validate the date for its continuation.

Validation against 1 is added directly to compare against numeric value 1 that can be assumed to be true.

The database all_day field in table c_calendar_event is set to 1 correctly.