LPs: Read lp_item extra field 'no_automatic_validation' see BT#18568
Chamilo by default changes the lp item to status = complete when the
LP is a document. Setting the "no_automatic_validation" to true,
Chamilo will not do that. The HTML file will be in charge of that update
The document can update the status with LMSCommit('iframe');

Example in the HTML:

<script>
    document.addEventListener("DOMContentLoaded", function(event) {
        if (window.parent && window.parent.API) {
            let api = window.parent.API;
            if (api) {
                api.LMSSetValue('cmi.core.score.raw', 100);
                api.LMSSetValue("cmi.core.lesson_status", "completed");
                api.LMSCommit('iframe');
            }
        }
    });
</script>