Make UserRelUserTest pass again
When a user A make a friend request to user B, and then user B accepts
the request, a new UserRelUser entity is created first with the
requested friend constant, and then updated to the friend constant.
Only one row is added, and therefore getFriends() method on user A will
return user B, but on user B it will not provide any value.
Instead when calling getFriendsWithMe() on user A it will provide no
values, but on user B it will return A.

Adjusted the test to reflect that.

Also, tweaked a bit for easier reading.